+++ activerecord/lib/active_record/type/value.rb
@@ -1,0 +1,5 @@
+module ActiveRecord
+  module Type
+    class Value < ActiveModel::Type::Value; end
+  end
+end