+++ activerecord/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Made ActiveRecord consistently use `ActiveRecord::Type` (not `ActiveModel::Type`)
+
+    *Iain Beeston*
+
 *   Serialize JSON attribute value `nil` as SQL `NULL`, not JSON `null`
 
     *Trung Duc Tran*